fix(commands.create_note): allow create/select subdirectory before creating a new note
This change will add a select prompt before note creation to select a
subdirectory or create a new subdirectory for the note creation.
Previously we can already just type `folder/something` and it will
create the folder. Now the behaviour are still the same, but we added a
new select picker for easier selection of subdirectory as we go.

Note that if a subdirectory is selected and the input also has
subdirectory in it, it will get nested. For example:

- Selected: `linux-note`
- Input: `linux-note/linux-is-amazing.md`
- Result: `linux-note/linux-note/linux-is-amazing.md`

It's recommended to just type in the name or filename instead of nested
path with filename