feat(commands): add `open` command (#40)
* feat(commands): add `open` command

This change adds another command to open the desired file slightly
quicker. Not sure how this will differs with the `pick` function, but
let's go with this first.

I think the biggest difference is that `open` will try to actually open
the most relevant file, where `pick` is being used when you do not
exactly know what to open but wants to view everything that you have
first.

* chore(build): auto-generate docs from readme

This is an automated commit.