feat(commands.open): add ability to `pluralised` the query
This changes will allow the behaviour of pluralising the query for
opening a file. For example: searching for `campaign` will try to match
`campaign` and also `campaigns`.

In my opinion, this changes is good for the `open` command, as
sometimes, we don't actually remember everything, especially whether it
is pluralised or not.