feat(commands.picker): add `snacks`, `telescope` and `fzf-lua` picker (#42)
* feat(commands.picker): add `snacks`, `telescope` and `fzf-lua` picker

With this change, user can now setup their desired `picker` in the
config, and when call the picker function, it will pick up the
configured picker from config.

User can also on demand select other picker in the `pick` function if
they want to.

Regardless what picker are configured, if the picker is not found or not
configured, it will fallback to the basic `vim.ui.select` picker.

* chore(build): auto-generate docs from readme

This is an automated commit.