fix(commands.create_todo_today): support rollover for multiple headings with configuration (#57)
* fix(commands.create_todo_today): support rollover for multiple headings with configuration

This change enables user to configure not just one heading but multiple
headings that to be rolled over if any. Note that it only supports `h2`
for now.

Rolled over todos will follow the previous todo's headings and group
them properly.

If the previous has some heading specified but doesn't exist in the
target, it will get appended at the bottom of the target with rollover
details if any.

If the config has some headers configured that doesn't exist in the
previous, it will get ignored.

* chore(build): auto-generate docs from readme

This is an automated commit.