fix(commands.navigate): respect `default_split` settings when using `navigate`
Before this, we are just using `edit` command, it works fine until the
introduction of `floating` splits. It will navigate on the main buffer
instead of the floating buffer.

This change will force the navigation to happen on the float buffer if
configured, or else on the main buffer without further spliting